+++ public/img/index.js
@@ -1,3 +1,4 @@
+require('./favicon.ico');
 require('./default-icon.png');
 require('./google_play_store_btn@2x.png');
 require('./ios_app_store_btn@2x.png');