+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ##### Bug Fixes
 * Allow user with event access to update recent event
+* Show default icon on web when creating an observation if no primary/secondary fields selected
 
 ## [5.1.0](https://github.com/ngageoint/mage-server/releases/tag/5.1.0) (02-27-2018)