+++ CHANGELOG.md
@@ -3,9 +3,15 @@ All notable changes to this project will be documented in this file.
 Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
-## 5.1.1 (TBD)
+## 5.1.2 (TBD)
 
 * TBD
+
+##### Features
+
+##### Bug Fixes
+
+## [5.1.1](https://github.com/ngageoint/mage-server/releases/tag/5.1.1) (04-12-2018)
 
 ##### Features
 * Added init.d script for mage application